\name{Figure6.1}
\alias{Figure6.1}
\docType{data}
\title{Figure6.1}
\usage{data("Figure6.1")}
\format{A data frame with 11 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Figure6.1)}
\keyword{{datasets}}
