\name{Problem10.1}
\alias{Problem10.1}
\docType{data}
\title{Exercise 10.1}
\usage{data("Problem10.1")}
\format{A data frame with 10 observations on the following variable(s).\describe{
\item{\code{Hardwood}}{a numeric vector}
\item{\code{Strength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem10.1)}
\keyword{{datasets}}
