\name{Problem10.11}
\alias{Problem10.11}
\docType{data}
\title{Exercise 10.11}
\usage{data("Problem10.11")}
\format{A data frame with 14 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{D}}{a numeric vector}
\item{\code{Rate}}{a numeric vector}
\item{\code{Ab}}{a numeric vector}
\item{\code{Ac}}{a numeric vector}
\item{\code{Ad}}{a numeric vector}
\item{\code{Bc}}{a numeric vector}
\item{\code{Bd}}{a numeric vector}
\item{\code{Cd}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem10.11)}
\keyword{{datasets}}
