\name{Problem10.12}
\alias{Problem10.12}
\docType{data}
\title{Exercise 10.12}
\usage{data("Problem10.12")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{X1}}{a numeric vector}
\item{\code{X2}}{a numeric vector}
\item{\code{Y}}{a numeric vector}
\item{\code{X1Squared}}{a numeric vector}
\item{\code{X2Squared}}{a numeric vector}
\item{\code{X1x2}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem10.12)}
\keyword{{datasets}}
