\name{Problem10.18a}
\alias{Problem10.18a}
\docType{data}
\title{Exercise 10.18a}
\usage{data("Problem10.18a")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{X1}}{a numeric vector}
\item{\code{X2}}{a numeric vector}
\item{\code{X3}}{a numeric vector}
\item{\code{X4}}{a numeric vector}
\item{\code{X5}}{a numeric vector}
\item{\code{X6}}{a numeric vector}
\item{\code{X7}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem10.18a)}
\keyword{{datasets}}
