\name{Problem10.2}
\alias{Problem10.2}
\docType{data}
\title{Exercise 10.2}
\usage{data("Problem10.2")}
\format{A data frame with 15 observations on the following variable(s).\describe{
\item{\code{Pollution}}{a numeric vector}
\item{\code{Purity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem10.2)}
\keyword{{datasets}}
