\name{Problem10.6}
\alias{Problem10.6}
\docType{data}
\title{Exercise 10.6}
\usage{data("Problem10.6")}
\format{A data frame with 6 observations on the following variable(s).\describe{
\item{\code{Viscosity}}{a numeric vector}
\item{\code{Load}}{a numeric vector}
\item{\code{Wear}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem10.6)}
\keyword{{datasets}}
