\name{Problem10.7}
\alias{Problem10.7}
\docType{data}
\title{Exercise 10.7}
\usage{data("Problem10.7")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{BrakeHorsepower}}{a numeric vector}
\item{\code{Rpm}}{a numeric vector}
\item{\code{RoadOctaneNumber}}{a numeric vector}
\item{\code{Compression}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem10.7)}
\keyword{{datasets}}
