\name{Problem10.9}
\alias{Problem10.9}
\docType{data}
\title{Exercise 10.9}
\usage{data("Problem10.9")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{Concentration}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem10.9)}
\keyword{{datasets}}
