\name{Problem11.1}
\alias{Problem11.1}
\docType{data}
\title{Exercise 11.1}
\usage{data("Problem11.1")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{Pressure}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Purity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.1)}
\keyword{{datasets}}
