\name{Problem11.10}
\alias{Problem11.10}
\docType{data}
\title{Exercise 11.10}
\usage{data("Problem11.10")}
\format{A data frame with 11 observations on the following variable(s).\describe{
\item{\code{X1}}{a numeric vector}
\item{\code{X2}}{a numeric vector}
\item{\code{Y}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.10)}
\keyword{{datasets}}
