\name{Problem11.11}
\alias{Problem11.11}
\docType{data}
\title{Exercise 11.11}
\usage{data("Problem11.11")}
\format{A data frame with 15 observations on the following variable(s).\describe{
\item{\code{Temperatue}}{a numeric vector}
\item{\code{AgitationRate}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Viscosity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.11)}
\keyword{{datasets}}
