\name{Problem11.12}
\alias{Problem11.12}
\docType{data}
\title{Exercise 11.12}
\usage{data("Problem11.12")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{Time}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Catalyst}}{a numeric vector}
\item{\code{Conversion}}{a numeric vector}
\item{\code{Activity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.12)}
\keyword{{datasets}}
