\name{Problem11.13}
\alias{Problem11.13}
\docType{data}
\title{Exercise 11.13}
\usage{data("Problem11.13")}
\format{A data frame with 4 observations on the following variable(s).\describe{
\item{\code{Hardness}}{a numeric vector}
\item{\code{Time}}{a numeric vector}
\item{\code{Life}}{a numeric vector}
\item{\code{Cost}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.13)}
\keyword{{datasets}}
