\name{Problem11.14}
\alias{Problem11.14}
\docType{data}
\title{Exercise 11.14}
\usage{data("Problem11.14")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{X1}}{a numeric vector}
\item{\code{X2}}{a numeric vector}
\item{\code{MeanThick}}{a numeric vector}
\item{\code{VarThick}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.14)}
\keyword{{datasets}}
