\name{Problem11.2}
\alias{Problem11.2}
\docType{data}
\title{Exercise 11.2}
\usage{data("Problem11.2")}
\format{A data frame with 11 observations on the following variable(s).\describe{
\item{\code{Item1Qty}}{a numeric vector}
\item{\code{Item1Reorder}}{a numeric vector}
\item{\code{Item2Qty}}{a numeric vector}
\item{\code{Item2Reorder}}{a numeric vector}
\item{\code{TotalCost}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.2)}
\keyword{{datasets}}
