\name{Problem11.31}
\alias{Problem11.31}
\docType{data}
\title{Exercise 11.31}
\usage{data("Problem11.31")}
\format{A data frame with 14 observations on the following variable(s).\describe{
\item{\code{X1}}{a numeric vector}
\item{\code{X2}}{a numeric vector}
\item{\code{X3}}{a numeric vector}
\item{\code{Y}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.31)}
\keyword{{datasets}}
