\name{Problem11.32}
\alias{Problem11.32}
\docType{data}
\title{Exercise 11.32}
\usage{data("Problem11.32")}
\format{A data frame with 13 observations on the following variable(s).\describe{
\item{\code{X1}}{a numeric vector}
\item{\code{X2}}{a numeric vector}
\item{\code{X3}}{a numeric vector}
\item{\code{X4}}{a numeric vector}
\item{\code{X5}}{a numeric vector}
\item{\code{X6}}{a numeric vector}
\item{\code{Y}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.32)}
\keyword{{datasets}}
