\name{Problem11.33}
\alias{Problem11.33}
\docType{data}
\title{Exercise 11.33}
\usage{data("Problem11.33")}
\format{A data frame with 19 observations on the following variable(s).\describe{
\item{\code{Run}}{a character vector}
\item{\code{Viscosity}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{PlateGap}}{a numeric vector}
\item{\code{Tear}}{a numeric vector}
\item{\code{Leakage}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.33)}
\keyword{{datasets}}
