\name{Problem11.34}
\alias{Problem11.34}
\docType{data}
\title{Exercise 11.34}
\usage{data("Problem11.34")}
\format{A data frame with 30 observations on the following variable(s).\describe{
\item{\code{StdOrd}}{a numeric vector}
\item{\code{RunOrd}}{a numeric vector}
\item{\code{WingArea}}{a numeric vector}
\item{\code{WingRatio}}{a numeric vector}
\item{\code{BaseWidth}}{a numeric vector}
\item{\code{BaseLength}}{a numeric vector}
\item{\code{AvgFlightTime}}{a numeric vector}
\item{\code{StDevFlightTime}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.34)}
\keyword{{datasets}}
