\name{Problem11.36}
\alias{Problem11.36}
\docType{data}
\title{Exercise 11.36}
\usage{data("Problem11.36")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{StdOrd}}{a numeric vector}
\item{\code{X1}}{a numeric vector}
\item{\code{X2}}{a numeric vector}
\item{\code{X3}}{a numeric vector}
\item{\code{Cef}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.36)}
\keyword{{datasets}}
