\name{Problem11.37}
\alias{Problem11.37}
\docType{data}
\title{Exercise 11.37}
\usage{data("Problem11.37")}
\format{A data frame with 27 observations on the following variable(s).\describe{
\item{\code{StdOrd}}{a numeric vector}
\item{\code{X1}}{a numeric vector}
\item{\code{X2}}{a numeric vector}
\item{\code{X3}}{a numeric vector}
\item{\code{X4}}{a numeric vector}
\item{\code{MtgaseActivity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.37)}
\keyword{{datasets}}
