\name{Problem11.9}
\alias{Problem11.9}
\docType{data}
\title{Exercise 11.9}
\usage{data("Problem11.9")}
\format{A data frame with 13 observations on the following variable(s).\describe{
\item{\code{Temperature}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Time}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem11.9)}
\keyword{{datasets}}
