\name{Problem12.1}
\alias{Problem12.1}
\docType{data}
\title{Exercise 12.1}
\usage{data("Problem12.1")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{FurnaceTemp}}{a numeric vector}
\item{\code{HeatingTime}}{a numeric vector}
\item{\code{TransferTime}}{a numeric vector}
\item{\code{HoldTime}}{a numeric vector}
\item{\code{FreeHeightMean}}{a numeric vector}
\item{\code{FreeHeightVariance}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem12.1)}
\keyword{{datasets}}
