\name{Problem12.10}
\alias{Problem12.10}
\docType{data}
\title{Exercise 12.10}
\usage{data("Problem12.10")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{A}}{a character vector}
\item{\code{B}}{a character vector}
\item{\code{C}}{a character vector}
\item{\code{D}}{a character vector}
\item{\code{CoEmis}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem12.10)}
\keyword{{datasets}}
