\name{Problem12.11}
\alias{Problem12.11}
\docType{data}
\title{Exercise 12.11}
\usage{data("Problem12.11")}
\format{A data frame with 27 observations on the following variable(s).\describe{
\item{\code{Speed}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Distance}}{a numeric vector}
\item{\code{Mean}}{a numeric vector}
\item{\code{Std.Dev.}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem12.11)}
\keyword{{datasets}}
