\name{Problem12.14}
\alias{Problem12.14}
\docType{data}
\title{Exercise 12.14}
\usage{data("Problem12.14")}
\format{A data frame with 29 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{D}}{a numeric vector}
\item{\code{E}}{a numeric vector}
\item{\code{F}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem12.14)}
\keyword{{datasets}}
