\name{Problem12.18}
\alias{Problem12.18}
\docType{data}
\title{Exercise 12.18}
\usage{data("Problem12.18")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{D}}{a numeric vector}
\item{\code{E}}{a numeric vector}
\item{\code{SolderDefectsMean}}{a numeric vector}
\item{\code{SolderDefectsVariance}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem12.18)}
\keyword{{datasets}}
