\name{Problem12.2}
\alias{Problem12.2}
\docType{data}
\title{Exercise 12.2}
\usage{data("Problem12.2")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Carbonation}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Speed}}{a numeric vector}
\item{\code{FillDeviation}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem12.2)}
\keyword{{datasets}}
