\name{Problem12.21}
\alias{Problem12.21}
\docType{data}
\title{Exercise 12.21}
\usage{data("Problem12.21")}
\format{A data frame with 32 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{D}}{a numeric vector}
\item{\code{CrackLength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem12.21)}
\keyword{{datasets}}
