\name{Problem12.8}
\alias{Problem12.8}
\docType{data}
\title{Exercise 12.8}
\usage{data("Problem12.8")}
\format{A data frame with 15 observations on the following variable(s).\describe{
\item{\code{Temperature}}{a numeric vector}
\item{\code{AgitationRate}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Viscosity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem12.8)}
\keyword{{datasets}}
