\name{Problem12.9}
\alias{Problem12.9}
\docType{data}
\title{Exercise 12.9}
\usage{data("Problem12.9")}
\format{A data frame with 4 observations on the following variable(s).\describe{
\item{\code{Concentration}}{a numeric vector}
\item{\code{StirRate}}{a numeric vector}
\item{\code{Mean}}{a numeric vector}
\item{\code{Variance}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem12.9)}
\keyword{{datasets}}
