\name{Problem13.11}
\alias{Problem13.11}
\docType{data}
\title{Exercise 13.11}
\usage{data("Problem13.11")}
\format{A data frame with 90 observations on the following variable(s).\describe{
\item{\code{Inspector}}{a numeric vector}
\item{\code{Test}}{a numeric vector}
\item{\code{Part}}{a numeric vector}
\item{\code{Impedance}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem13.11)}
\keyword{{datasets}}
