\name{Problem13.15}
\alias{Problem13.15}
\docType{data}
\title{Exercise 13.15}
\usage{data("Problem13.15")}
\format{A data frame with 72 observations on the following variable(s).\describe{
\item{\code{Temp}}{a numeric vector}
\item{\code{Operator}}{a numeric vector}
\item{\code{Gauge}}{a numeric vector}
\item{\code{Drop}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem13.15)}
\keyword{{datasets}}
