\name{Problem13.19}
\alias{Problem13.19}
\docType{data}
\title{Exercise 13.19}
\usage{data("Problem13.19")}
\format{A data frame with 54 observations on the following variable(s).\describe{
\item{\code{CycleTime}}{a character vector}
\item{\code{Operator}}{a character vector}
\item{\code{Temperature}}{a character vector}
\item{\code{Score}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem13.19)}
\keyword{{datasets}}
