\name{Problem13.26}
\alias{Problem13.26}
\docType{data}
\title{Exercise 13.26}
\usage{data("Problem13.26")}
\format{A data frame with 60 observations on the following variable(s).\describe{
\item{\code{Part}}{a numeric vector}
\item{\code{Operator}}{a numeric vector}
\item{\code{Measurement}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem13.26)}
\keyword{{datasets}}
