\name{Problem13.3}
\alias{Problem13.3}
\docType{data}
\title{Exercise 13.3}
\usage{data("Problem13.3")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Operator}}{a numeric vector}
\item{\code{Machine}}{a numeric vector}
\item{\code{Strength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem13.3)}
\keyword{{datasets}}
