\name{Problem13.4}
\alias{Problem13.4}
\docType{data}
\title{Exercise 13.4}
\usage{data("Problem13.4")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Row}}{a numeric vector}
\item{\code{Column}}{a numeric vector}
\item{\code{Response}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem13.4)}
\keyword{{datasets}}
