\name{Problem13.7}
\alias{Problem13.7}
\docType{data}
\title{Exercise 13.7}
\usage{data("Problem13.7")}
\format{A data frame with 90 observations on the following variable(s).\describe{
\item{\code{Inspector}}{a numeric vector}
\item{\code{Test}}{a numeric vector}
\item{\code{Part}}{a numeric vector}
\item{\code{Impedance}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem13.7)}
\keyword{{datasets}}
