\name{Problem13.9}
\alias{Problem13.9}
\docType{data}
\title{Exercise 13.9}
\usage{data("Problem13.9")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{Position}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Density}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem13.9)}
\keyword{{datasets}}
