\name{Problem14.1}
\alias{Problem14.1}
\docType{data}
\title{Exercise 14.1}
\usage{data("Problem14.1")}
\format{A data frame with 36 observations on the following variable(s).\describe{
\item{\code{BurnRate}}{a numeric vector}
\item{\code{Process}}{a numeric vector}
\item{\code{Batch}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.1)}
\keyword{{datasets}}
