\name{Problem14.11}
\alias{Problem14.11}
\docType{data}
\title{Exercise 14.11}
\usage{data("Problem14.11")}
\format{A data frame with 15 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{Y}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.11)}
\keyword{{datasets}}
