\name{Problem14.15}
\alias{Problem14.15}
\docType{data}
\title{Exercise 14.15}
\usage{data("Problem14.15")}
\format{A data frame with 54 observations on the following variable(s).\describe{
\item{\code{Machine}}{a numeric vector}
\item{\code{Power}}{a numeric vector}
\item{\code{Station}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.15)}
\keyword{{datasets}}
