\name{Problem14.16}
\alias{Problem14.16}
\docType{data}
\title{Exercise 14.16}
\usage{data("Problem14.16")}
\format{A data frame with 54 observations on the following variable(s).\describe{
\item{\code{Vendor}}{a numeric vector}
\item{\code{Heat}}{a numeric vector}
\item{\code{BarSize}}{a numeric vector}
\item{\code{Strength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.16)}
\keyword{{datasets}}
