\name{Problem14.19}
\alias{Problem14.19}
\docType{data}
\title{Exercise 14.19}
\usage{data("Problem14.19")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Temperature}}{a numeric vector}
\item{\code{Shift}}{a numeric vector}
\item{\code{Time}}{a numeric vector}
\item{\code{Strength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.19)}
\keyword{{datasets}}
