\name{Problem14.2}
\alias{Problem14.2}
\docType{data}
\title{Exercise 14.2}
\usage{data("Problem14.2")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Finish}}{a numeric vector}
\item{\code{Machine}}{a numeric vector}
\item{\code{Operator}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.2)}
\keyword{{datasets}}
