\name{Problem14.20}
\alias{Problem14.20}
\docType{data}
\title{Exercise 14.20}
\usage{data("Problem14.20")}
\format{A data frame with 36 observations on the following variable(s).\describe{
\item{\code{Mix}}{a numeric vector}
\item{\code{Method}}{a numeric vector}
\item{\code{Day}}{a numeric vector}
\item{\code{Reflectance}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.20)}
\keyword{{datasets}}
