\name{Problem14.22}
\alias{Problem14.22}
\docType{data}
\title{Exercise 14.22}
\usage{data("Problem14.22")}
\format{A data frame with 144 observations on the following variable(s).\describe{
\item{\code{Day}}{a numeric vector}
\item{\code{Tech}}{a numeric vector}
\item{\code{Dose}}{a numeric vector}
\item{\code{Thick}}{a numeric vector}
\item{\code{Time}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.22)}
\keyword{{datasets}}
