\name{Problem14.27}
\alias{Problem14.27}
\docType{data}
\title{Exercise 14.27}
\usage{data("Problem14.27")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Pressure}}{a numeric vector}
\item{\code{Power}}{a numeric vector}
\item{\code{GasFlow}}{a numeric vector}
\item{\code{GasType}}{a character vector}
\item{\code{PaperType}}{a character vector}
\item{\code{ContactAngle}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.27)}
\keyword{{datasets}}
