\name{Problem14.3}
\alias{Problem14.3}
\docType{data}
\title{Exercise 14.3}
\usage{data("Problem14.3")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Variability}}{a numeric vector}
\item{\code{Machine}}{a numeric vector}
\item{\code{Spindle}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.3)}
\keyword{{datasets}}
