\name{Problem14.4}
\alias{Problem14.4}
\docType{data}
\title{Exercise 14.4}
\usage{data("Problem14.4")}
\format{A data frame with 36 observations on the following variable(s).\describe{
\item{\code{Time}}{a numeric vector}
\item{\code{Job}}{a numeric vector}
\item{\code{Operator}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.4)}
\keyword{{datasets}}
