\name{Problem14.5}
\alias{Problem14.5}
\docType{data}
\title{Exercise 14.5}
\usage{data("Problem14.5")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Hardness}}{a numeric vector}
\item{\code{Alloy}}{a numeric vector}
\item{\code{Heat}}{a numeric vector}
\item{\code{Ingot}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.5)}
\keyword{{datasets}}
