\name{Problem14.7}
\alias{Problem14.7}
\docType{data}
\title{Exercise 14.7}
\usage{data("Problem14.7")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{Y}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem14.7)}
\keyword{{datasets}}
