\name{Problem15.10}
\alias{Problem15.10}
\docType{data}
\title{Exercise 15.10}
\usage{data("Problem15.10")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{SodiumCitrate}}{a numeric vector}
\item{\code{Glycerol}}{a numeric vector}
\item{\code{EquilibriumTime}}{a numeric vector}
\item{\code{NumberSurvived}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem15.10)}
\keyword{{datasets}}
