\name{Problem15.11}
\alias{Problem15.11}
\docType{data}
\title{Exercise 15.11}
\usage{data("Problem15.11")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Truck}}{a numeric vector}
\item{\code{Volume}}{a numeric vector}
\item{\code{Time}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem15.11)}
\keyword{{datasets}}
