\name{Problem15.15}
\alias{Problem15.15}
\docType{data}
\title{Exercise 15.15}
\usage{data("Problem15.15")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{Thick}}{a numeric vector}
\item{\code{Glue}}{a numeric vector}
\item{\code{Strength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem15.15)}
\keyword{{datasets}}
