\name{Problem15.17}
\alias{Problem15.17}
\docType{data}
\title{Exercise 15.17}
\usage{data("Problem15.17")}
\format{A data frame with 15 observations on the following variable(s).\describe{
\item{\code{Hardness}}{a numeric vector}
\item{\code{Speed}}{a numeric vector}
\item{\code{Removal}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem15.17)}
\keyword{{datasets}}
