\name{Problem15.2}
\alias{Problem15.2}
\docType{data}
\title{Exercise 15.2}
\usage{data("Problem15.2")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Load}}{a numeric vector}
\item{\code{Rate}}{a numeric vector}
\item{\code{Speed}}{a numeric vector}
\item{\code{Mud}}{a numeric vector}
\item{\code{AdvanceRate}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem15.2)}
\keyword{{datasets}}
