\name{Problem15.21}
\alias{Problem15.21}
\docType{data}
\title{Exercise 15.21}
\usage{data("Problem15.21")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Judge}}{a character vector}
\item{\code{Wine}}{a character vector}
\item{\code{WineQuality}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem15.21)}
\keyword{{datasets}}
