\name{Problem2.22}
\alias{Problem2.22}
\docType{data}
\title{Exercise 2.22}
\usage{data("Problem2.22")}
\format{A data frame with 10 observations on the following variable(s).\describe{
\item{\code{Days}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.22)}
\keyword{{datasets}}
