\name{Problem2.24}
\alias{Problem2.24}
\docType{data}
\title{Exercise 2.24}
\usage{data("Problem2.24")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Hours}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.24)}
\keyword{{datasets}}
