\name{Problem2.26}
\alias{Problem2.26}
\docType{data}
\title{Exercise 2.26}
\usage{data("Problem2.26")}
\format{A data frame with 10 observations on the following variable(s).\describe{
\item{\code{Machine1}}{a numeric vector}
\item{\code{Machine2}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.26)}
\keyword{{datasets}}
