\name{Problem2.29}
\alias{Problem2.29}
\docType{data}
\title{Exercise 2.29}
\usage{data("Problem2.29")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{FlowRate}}{a numeric vector}
\item{\code{Uniformity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.29)}
\keyword{{datasets}}
