\name{Problem2.31}
\alias{Problem2.31}
\docType{data}
\title{Exercise 2.31}
\usage{data("Problem2.31")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{Thinner}}{a numeric vector}
\item{\code{Thicker}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.31)}
\keyword{{datasets}}
