\name{Problem2.32}
\alias{Problem2.32}
\docType{data}
\title{Exercise 2.32}
\usage{data("Problem2.32")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{Faster}}{a numeric vector}
\item{\code{Slower}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.32)}
\keyword{{datasets}}
