\name{Problem2.33}
\alias{Problem2.33}
\docType{data}
\title{Exercise 2.33}
\usage{data("Problem2.33")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{Uniformity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.33)}
\keyword{{datasets}}
