\name{Problem2.34}
\alias{Problem2.34}
\docType{data}
\title{Exercise 2.34}
\usage{data("Problem2.34")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Caliper1}}{a numeric vector}
\item{\code{Caliper2}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.34)}
\keyword{{datasets}}
