\name{Problem2.35}
\alias{Problem2.35}
\docType{data}
\title{Exercise 2.35}
\usage{data("Problem2.35")}
\format{A data frame with 10 observations on the following variable(s).\describe{
\item{\code{Pair}}{a numeric vector}
\item{\code{Firstborn}}{a numeric vector}
\item{\code{Secondborn}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.35)}
\keyword{{datasets}}
