\name{Problem2.36}
\alias{Problem2.36}
\docType{data}
\title{Exercise 2.36}
\usage{data("Problem2.36")}
\format{A data frame with 9 observations on the following variable(s).\describe{
\item{\code{Karlsruhe}}{a numeric vector}
\item{\code{Lehigh}}{a numeric vector}
\item{\code{Difference}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.36)}
\keyword{{datasets}}
