\name{Problem2.37}
\alias{Problem2.37}
\docType{data}
\title{Exercise 2.37}
\usage{data("Problem2.37")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Form1}}{a numeric vector}
\item{\code{Form2}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.37)}
\keyword{{datasets}}
