\name{Problem2.39}
\alias{Problem2.39}
\docType{data}
\title{Exercise 2.39}
\usage{data("Problem2.39")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{Solution1}}{a numeric vector}
\item{\code{Solution2}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.39)}
\keyword{{datasets}}
