\name{Problem2.45}
\alias{Problem2.45}
\docType{data}
\title{Exercise 2.45}
\usage{data("Problem2.45")}
\format{A data frame with 10 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{Delta}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.45)}
\keyword{{datasets}}
