\name{Problem2.46}
\alias{Problem2.46}
\docType{data}
\title{Exercise 2.46}
\usage{data("Problem2.46")}
\format{A data frame with 10 observations on the following variable(s).\describe{
\item{\code{Type1}}{a numeric vector}
\item{\code{Type2}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem2.46)}
\keyword{{datasets}}
