\name{Problem3.10}
\alias{Problem3.10}
\docType{data}
\title{Exercise 3.10}
\usage{data("Problem3.10")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{Sham}}{a numeric vector}
\item{\code{OneHour}}{a numeric vector}
\item{\code{TwoHours}}{a numeric vector}
\item{\code{FourHours}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.10)}
\keyword{{datasets}}
