\name{Problem3.11}
\alias{Problem3.11}
\docType{data}
\title{Exercise 3.11}
\usage{data("Problem3.11")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Technique}}{a character vector}
\item{\code{TensileStrength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.11)}
\keyword{{datasets}}
