\name{Problem3.14}
\alias{Problem3.14}
\docType{data}
\title{Exercise 3.14}
\usage{data("Problem3.14")}
\format{A data frame with 25 observations on the following variable(s).\describe{
\item{\code{CottonWeightPct}}{a numeric vector}
\item{\code{Observations}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.14)}
\keyword{{datasets}}
