\name{Problem3.16}
\alias{Problem3.16}
\docType{data}
\title{Exercise 3.16}
\usage{data("Problem3.16")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Dosage}}{a character vector}
\item{\code{Activity}}{a numeric vector}
\item{\code{Resi1}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.16)}
\keyword{{datasets}}
