\name{Problem3.17}
\alias{Problem3.17}
\docType{data}
\title{Exercise 3.17}
\usage{data("Problem3.17")}
\format{A data frame with 40 observations on the following variable(s).\describe{
\item{\code{TypeOfCar}}{a character vector}
\item{\code{Days}}{a numeric vector}
\item{\code{SqrtDays}}{a numeric vector}
\item{\code{Resi1}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.17)}
\keyword{{datasets}}
