\name{Problem3.18}
\alias{Problem3.18}
\docType{data}
\title{Exercise 3.18}
\usage{data("Problem3.18")}
\format{A data frame with 25 observations on the following variable(s).\describe{
\item{\code{Season}}{a character vector}
\item{\code{Score}}{a numeric vector}
\item{\code{Resi1}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.18)}
\keyword{{datasets}}
