\name{Problem3.19}
\alias{Problem3.19}
\docType{data}
\title{Exercise 3.19}
\usage{data("Problem3.19")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Approach}}{a numeric vector}
\item{\code{Contribution}}{a numeric vector}
\item{\code{Resi1}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.19)}
\keyword{{datasets}}
