\name{Problem3.20}
\alias{Problem3.20}
\docType{data}
\title{Exercise 3.20}
\usage{data("Problem3.20")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{Temp}}{a numeric vector}
\item{\code{Density}}{a numeric vector}
\item{\code{Resi1}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.20)}
\keyword{{datasets}}
