\name{Problem3.22}
\alias{Problem3.22}
\docType{data}
\title{Exercise 3.22}
\usage{data("Problem3.22")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{CoatingType}}{a character vector}
\item{\code{Conductivity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.22)}
\keyword{{datasets}}
