\name{Problem3.24}
\alias{Problem3.24}
\docType{data}
\title{Exercise 3.24}
\usage{data("Problem3.24")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{RoddingLevel}}{a character vector}
\item{\code{Strength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.24)}
\keyword{{datasets}}
