\name{Problem3.25}
\alias{Problem3.25}
\docType{data}
\title{Exercise 3.25}
\usage{data("Problem3.25")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{OrificeDiameter}}{a character vector}
\item{\code{RadonReleased}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.25)}
\keyword{{datasets}}
