\name{Problem3.26}
\alias{Problem3.26}
\docType{data}
\title{Exercise 3.26}
\usage{data("Problem3.26")}
\format{A data frame with 15 observations on the following variable(s).\describe{
\item{\code{CircuitType}}{a character vector}
\item{\code{ResponseTime}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.26)}
\keyword{{datasets}}
