\name{Problem3.27}
\alias{Problem3.27}
\docType{data}
\title{Exercise 3.27}
\usage{data("Problem3.27")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Catalyst}}{a character vector}
\item{\code{Concentration}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.27)}
\keyword{{datasets}}
