\name{Problem3.28}
\alias{Problem3.28}
\docType{data}
\title{Exercise 3.28}
\usage{data("Problem3.28")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{Material}}{a character vector}
\item{\code{FailureTime}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.28)}
\keyword{{datasets}}
