\name{Problem3.29}
\alias{Problem3.29}
\docType{data}
\title{Exercise 3.29}
\usage{data("Problem3.29")}
\format{A data frame with 15 observations on the following variable(s).\describe{
\item{\code{Method}}{a character vector}
\item{\code{Count}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.29)}
\keyword{{datasets}}
