\name{Problem3.31}
\alias{Problem3.31}
\docType{data}
\title{Exercise 3.31}
\usage{data("Problem3.31")}
\format{A data frame with 15 observations on the following variable(s).\describe{
\item{\code{Oven}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.31)}
\keyword{{datasets}}
