\name{Problem3.32}
\alias{Problem3.32}
\docType{data}
\title{Exercise 3.32}
\usage{data("Problem3.32")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{WaferPosition}}{a numeric vector}
\item{\code{Uniformity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.32)}
\keyword{{datasets}}
