\name{Problem3.34}
\alias{Problem3.34}
\docType{data}
\title{Exercise 3.34}
\usage{data("Problem3.34")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{Chemical}}{a numeric vector}
\item{\code{Brightness}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.34)}
\keyword{{datasets}}
