\name{Problem3.40}
\alias{Problem3.40}
\docType{data}
\title{Exercise 3.40}
\usage{data("Problem3.40")}
\format{A data frame with 15 observations on the following variable(s).\describe{
\item{\code{Brand}}{a character vector}
\item{\code{Deviation}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.40)}
\keyword{{datasets}}
