\name{Problem3.46}
\alias{Problem3.46}
\docType{data}
\title{Exercise 3.46}
\usage{data("Problem3.46")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Algorithm}}{a character vector}
\item{\code{CellAverage}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.46)}
\keyword{{datasets}}
