\name{Problem3.47}
\alias{Problem3.47}
\docType{data}
\title{Exercise 3.47}
\usage{data("Problem3.47")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Algorithm}}{a character vector}
\item{\code{CellStdev}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.47)}
\keyword{{datasets}}
