\name{Problem3.48}
\alias{Problem3.48}
\docType{data}
\title{Exercise 3.48}
\usage{data("Problem3.48")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{FeedRate}}{a character vector}
\item{\code{RunStdev}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.48)}
\keyword{{datasets}}
