\name{Problem3.53}
\alias{Problem3.53}
\docType{data}
\title{Exercise 3.53}
\usage{data("Problem3.53")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{Power}}{a numeric vector}
\item{\code{EtchRate}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.53)}
\keyword{{datasets}}
