\name{Problem3.54}
\alias{Problem3.54}
\docType{data}
\title{Exercise 3.54}
\usage{data("Problem3.54")}
\format{A data frame with 25 observations on the following variable(s).\describe{
\item{\code{Loom}}{a character vector}
\item{\code{Output}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.54)}
\keyword{{datasets}}
