\name{Problem3.55}
\alias{Problem3.55}
\docType{data}
\title{Exercise 3.55}
\usage{data("Problem3.55")}
\format{A data frame with 25 observations on the following variable(s).\describe{
\item{\code{Batch}}{a character vector}
\item{\code{CalciumContent}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem3.55)}
\keyword{{datasets}}
