\name{Problem4.12}
\alias{Problem4.12}
\docType{data}
\title{Exercise 4.12}
\usage{data("Problem4.12")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Tip}}{a character vector}
\item{\code{Coupon}}{a numeric vector}
\item{\code{Hardness}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.12)}
\keyword{{datasets}}
