\name{Problem4.14}
\alias{Problem4.14}
\docType{data}
\title{Exercise 4.14}
\usage{data("Problem4.14")}
\format{A data frame with 15 observations on the following variable(s).\describe{
\item{\code{Oil}}{a character vector}
\item{\code{Truck}}{a numeric vector}
\item{\code{FuelConsumption}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.14)}
\keyword{{datasets}}
