\name{Problem4.15}
\alias{Problem4.15}
\docType{data}
\title{Exercise 4.15}
\usage{data("Problem4.15")}
\format{A data frame with 30 observations on the following variable(s).\describe{
\item{\code{NozzleDesign}}{a character vector}
\item{\code{JetVelocity}}{a numeric vector}
\item{\code{Shape}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.15)}
\keyword{{datasets}}
