\name{Problem4.16}
\alias{Problem4.16}
\docType{data}
\title{Exercise 4.16}
\usage{data("Problem4.16")}
\format{A data frame with 36 observations on the following variable(s).\describe{
\item{\code{Project}}{a numeric vector}
\item{\code{Algorith}}{a character vector}
\item{\code{CostError}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.16)}
\keyword{{datasets}}
