\name{Problem4.17}
\alias{Problem4.17}
\docType{data}
\title{Exercise 4.17}
\usage{data("Problem4.17")}
\format{A data frame with 30 observations on the following variable(s).\describe{
\item{\code{Project}}{a numeric vector}
\item{\code{Treatment}}{a character vector}
\item{\code{GeneExpression}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.17)}
\keyword{{datasets}}
