\name{Problem4.18}
\alias{Problem4.18}
\docType{data}
\title{Exercise 4.18}
\usage{data("Problem4.18")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Algorithm}}{a character vector}
\item{\code{TimePeriod}}{a numeric vector}
\item{\code{Average}}{a numeric vector}
\item{\code{Stdev}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.18)}
\keyword{{datasets}}
