\name{Problem4.19}
\alias{Problem4.19}
\docType{data}
\title{Exercise 4.19}
\usage{data("Problem4.19")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{StirringRate}}{a character vector}
\item{\code{Furnace}}{a numeric vector}
\item{\code{GrainSize}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.19)}
\keyword{{datasets}}
