\name{Problem4.26}
\alias{Problem4.26}
\docType{data}
\title{Exercise 4.26}
\usage{data("Problem4.26")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{Distance}}{a character vector}
\item{\code{Subject}}{a numeric vector}
\item{\code{FocusTime}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.26)}
\keyword{{datasets}}
