\name{Problem4.27}
\alias{Problem4.27}
\docType{data}
\title{Exercise 4.27}
\usage{data("Problem4.27")}
\format{A data frame with 25 observations on the following variable(s).\describe{
\item{\code{Batch}}{a numeric vector}
\item{\code{Day}}{a numeric vector}
\item{\code{Catalyst}}{a character vector}
\item{\code{Time}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.27)}
\keyword{{datasets}}
