\name{Problem4.28}
\alias{Problem4.28}
\docType{data}
\title{Exercise 4.28}
\usage{data("Problem4.28")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Order}}{a numeric vector}
\item{\code{Operator}}{a numeric vector}
\item{\code{Method}}{a character vector}
\item{\code{Time}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.28)}
\keyword{{datasets}}
