\name{Problem4.40}
\alias{Problem4.40}
\docType{data}
\title{Exercise 4.40}
\usage{data("Problem4.40")}
\format{A data frame with 25 observations on the following variable(s).\describe{
\item{\code{Batch}}{a numeric vector}
\item{\code{Acid}}{a numeric vector}
\item{\code{Time}}{a character vector}
\item{\code{Catalyst}}{a character vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.40)}
\keyword{{datasets}}
