\name{Problem4.43}
\alias{Problem4.43}
\docType{data}
\title{Exercise 4.43}
\usage{data("Problem4.43")}
\format{A data frame with 32 observations on the following variable(s).\describe{
\item{\code{Order}}{a numeric vector}
\item{\code{Operator}}{a numeric vector}
\item{\code{Method}}{a character vector}
\item{\code{Time}}{a numeric vector}
\item{\code{Square}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.43)}
\keyword{{datasets}}
