\name{Problem4.45}
\alias{Problem4.45}
\docType{data}
\title{Exercise 4.45}
\usage{data("Problem4.45")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{Additive}}{a numeric vector}
\item{\code{Car}}{a numeric vector}
\item{\code{Mileage}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.45)}
\keyword{{datasets}}
