\name{Problem4.47}
\alias{Problem4.47}
\docType{data}
\title{Exercise 4.47}
\usage{data("Problem4.47")}
\format{A data frame with 21 observations on the following variable(s).\describe{
\item{\code{Concentration}}{a numeric vector}
\item{\code{Days}}{a numeric vector}
\item{\code{Strength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.47)}
\keyword{{datasets}}
