\name{Problem4.59}
\alias{Problem4.59}
\docType{data}
\title{Exercise 4.59}
\usage{data("Problem4.59")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Design}}{a character vector}
\item{\code{Region}}{a character vector}
\item{\code{NumberOfResponses}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.59)}
\keyword{{datasets}}
