\name{Problem4.8}
\alias{Problem4.8}
\docType{data}
\title{Exercise 4.8}
\usage{data("Problem4.8")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{Chemist}}{a character vector}
\item{\code{Bolt}}{a numeric vector}
\item{\code{Strength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.8)}
\keyword{{datasets}}
