\name{Problem4.9}
\alias{Problem4.9}
\docType{data}
\title{Exercise 4.9}
\usage{data("Problem4.9")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Solution}}{a character vector}
\item{\code{Days}}{a numeric vector}
\item{\code{Growth}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem4.9)}
\keyword{{datasets}}
