\name{Problem5.11}
\alias{Problem5.11}
\docType{data}
\title{Exercise 5.11}
\usage{data("Problem5.11")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{PhosphorousType}}{a numeric vector}
\item{\code{GlassType}}{a numeric vector}
\item{\code{Current}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.11)}
\keyword{{datasets}}
