\name{Problem5.12}
\alias{Problem5.12}
\docType{data}
\title{Exercise 5.12}
\usage{data("Problem5.12")}
\format{A data frame with 32 observations on the following variable(s).\describe{
\item{\code{CopperContent}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Warping}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.12)}
\keyword{{datasets}}
