\name{Problem5.13}
\alias{Problem5.13}
\docType{data}
\title{Exercise 5.13}
\usage{data("Problem5.13")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Operator}}{a numeric vector}
\item{\code{Machine}}{a numeric vector}
\item{\code{Strength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.13)}
\keyword{{datasets}}
