\name{Problem5.14}
\alias{Problem5.14}
\docType{data}
\title{Exercise 5.14}
\usage{data("Problem5.14")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{DrillSpeed}}{a numeric vector}
\item{\code{FeedRate}}{a numeric vector}
\item{\code{Force}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.14)}
\keyword{{datasets}}
