\name{Problem5.15}
\alias{Problem5.15}
\docType{data}
\title{Exercise 5.15}
\usage{data("Problem5.15")}
\format{A data frame with 27 observations on the following variable(s).\describe{
\item{\code{GlassType}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{LightOutput}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.15)}
\keyword{{datasets}}
