\name{Problem5.18}
\alias{Problem5.18}
\docType{data}
\title{Exercise 5.18}
\usage{data("Problem5.18")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{Position}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Density}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.18)}
\keyword{{datasets}}
