\name{Problem5.20}
\alias{Problem5.20}
\docType{data}
\title{Exercise 5.20}
\usage{data("Problem5.20")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Row}}{a numeric vector}
\item{\code{Column}}{a numeric vector}
\item{\code{Data}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.20)}
\keyword{{datasets}}
