\name{Problem5.21}
\alias{Problem5.21}
\docType{data}
\title{Exercise 5.21}
\usage{data("Problem5.21")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Pressure}}{a character vector}
\item{\code{Temperature}}{a character vector}
\item{\code{Strength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.21)}
\keyword{{datasets}}
