\name{Problem5.23}
\alias{Problem5.23}
\docType{data}
\title{Exercise 5.23}
\usage{data("Problem5.23")}
\format{A data frame with 36 observations on the following variable(s).\describe{
\item{\code{Hardwood}}{a numeric vector}
\item{\code{CookingTime}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Strength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.23)}
\keyword{{datasets}}
