\name{Problem5.26}
\alias{Problem5.26}
\docType{data}
\title{Exercise 5.26}
\usage{data("Problem5.26")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{Pressure}}{a character vector}
\item{\code{Temperature}}{a character vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.26)}
\keyword{{datasets}}
