\name{Problem5.28}
\alias{Problem5.28}
\docType{data}
\title{Exercise 5.28}
\usage{data("Problem5.28")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Operator}}{a character vector}
\item{\code{Machine}}{a character vector}
\item{\code{Strength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.28)}
\keyword{{datasets}}
