\name{Problem5.29}
\alias{Problem5.29}
\docType{data}
\title{Exercise 5.29}
\usage{data("Problem5.29")}
\format{A data frame with 36 observations on the following variable(s).\describe{
\item{\code{Frequency}}{a character vector}
\item{\code{Environment}}{a character vector}
\item{\code{CrackGrowth}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.29)}
\keyword{{datasets}}
