\name{Problem5.30}
\alias{Problem5.30}
\docType{data}
\title{Exercise 5.30}
\usage{data("Problem5.30")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Doping}}{a character vector}
\item{\code{Anneal}}{a character vector}
\item{\code{BaseCurrent}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.30)}
\keyword{{datasets}}
