\name{Problem5.31}
\alias{Problem5.31}
\docType{data}
\title{Exercise 5.31}
\usage{data("Problem5.31")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Battery}}{a character vector}
\item{\code{Device}}{a character vector}
\item{\code{LifeHours}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.31)}
\keyword{{datasets}}
