\name{Problem5.32}
\alias{Problem5.32}
\docType{data}
\title{Exercise 5.32}
\usage{data("Problem5.32")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{Clubs}}{a character vector}
\item{\code{Course}}{a character vector}
\item{\code{Score}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.32)}
\keyword{{datasets}}
