\name{Problem5.33}
\alias{Problem5.33}
\docType{data}
\title{Exercise 5.33}
\usage{data("Problem5.33")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Formulation}}{a character vector}
\item{\code{Washings}}{a character vector}
\item{\code{Booster}}{a character vector}
\item{\code{StainShade}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.33)}
\keyword{{datasets}}
