\name{Problem5.35}
\alias{Problem5.35}
\docType{data}
\title{Exercise 5.35}
\usage{data("Problem5.35")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{KeySize}}{a character vector}
\item{\code{KeyboardFeel}}{a character vector}
\item{\code{Speed}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.35)}
\keyword{{datasets}}
