\name{Problem5.36}
\alias{Problem5.36}
\docType{data}
\title{Exercise 5.36}
\usage{data("Problem5.36")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{AmmoniumPct}}{a numeric vector}
\item{\code{StirRate}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Density}}{a numeric vector}
\item{\code{SurfaceArea}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.36)}
\keyword{{datasets}}
