\name{Problem5.38}
\alias{Problem5.38}
\docType{data}
\title{Exercise 5.38}
\usage{data("Problem5.38")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{ExtractionPct}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.38)}
\keyword{{datasets}}
