\name{Problem5.40}
\alias{Problem5.40}
\docType{data}
\title{Exercise 5.40}
\usage{data("Problem5.40")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{PhosphorType}}{a numeric vector}
\item{\code{GlassType}}{a numeric vector}
\item{\code{Current}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.40)}
\keyword{{datasets}}
