\name{Problem5.43}
\alias{Problem5.43}
\docType{data}
\title{Exercise 5.43}
\usage{data("Problem5.43")}
\format{A data frame with 54 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{CycleTime}}{a numeric vector}
\item{\code{Operator}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Score}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.43)}
\keyword{{datasets}}
