\name{Problem5.44}
\alias{Problem5.44}
\docType{data}
\title{Exercise 5.44}
\usage{data("Problem5.44")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{AnchorType}}{a character vector}
\item{\code{FoamDensity}}{a character vector}
\item{\code{Force}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.44)}
\keyword{{datasets}}
