\name{Problem5.45}
\alias{Problem5.45}
\docType{data}
\title{Exercise 5.45}
\usage{data("Problem5.45")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{KeySize}}{a character vector}
\item{\code{KeyboardFeel}}{a character vector}
\item{\code{Speed}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.45)}
\keyword{{datasets}}
