\name{Problem5.8}
\alias{Problem5.8}
\docType{data}
\title{Exercise 5.8}
\usage{data("Problem5.8")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{Pressure}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.8)}
\keyword{{datasets}}
