\name{Problem5.9}
\alias{Problem5.9}
\docType{data}
\title{Exercise 5.9}
\usage{data("Problem5.9")}
\format{A data frame with 36 observations on the following variable(s).\describe{
\item{\code{DepthOfCut}}{a numeric vector}
\item{\code{FeedRate}}{a numeric vector}
\item{\code{SurfaceFinish}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem5.9)}
\keyword{{datasets}}
