\name{Problem6.10}
\alias{Problem6.10}
\docType{data}
\title{Exercise 6.10}
\usage{data("Problem6.10")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{CuttingSpeed}}{a numeric vector}
\item{\code{ToolGeometry}}{a numeric vector}
\item{\code{CuttingAngle}}{a numeric vector}
\item{\code{LifeHours}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.10)}
\keyword{{datasets}}
