\name{Problem6.12}
\alias{Problem6.12}
\docType{data}
\title{Exercise 6.12}
\usage{data("Problem6.12")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{CultureMedium}}{a character vector}
\item{\code{Time}}{a numeric vector}
\item{\code{VirusGrowth}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.12)}
\keyword{{datasets}}
