\name{Problem6.14}
\alias{Problem6.14}
\docType{data}
\title{Exercise 6.14}
\usage{data("Problem6.14")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Worker}}{a character vector}
\item{\code{BottleType}}{a character vector}
\item{\code{Pulse}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.14)}
\keyword{{datasets}}
