\name{Problem6.16}
\alias{Problem6.16}
\docType{data}
\title{Exercise 6.16}
\usage{data("Problem6.16")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{FlowRate}}{a numeric vector}
\item{\code{DepTime}}{a numeric vector}
\item{\code{Thickness}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.16)}
\keyword{{datasets}}
