\name{Problem6.19}
\alias{Problem6.19}
\docType{data}
\title{Exercise 6.19}
\usage{data("Problem6.19")}
\format{A data frame with 32 observations on the following variable(s).\describe{
\item{\code{PourTemperature}}{a numeric vector}
\item{\code{TitaniumContent}}{a numeric vector}
\item{\code{HeatTreatMethod}}{a character vector}
\item{\code{GrainRefiner}}{a numeric vector}
\item{\code{CrackLength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.19)}
\keyword{{datasets}}
