\name{Problem6.21}
\alias{Problem6.21}
\docType{data}
\title{Exercise 6.21}
\usage{data("Problem6.21")}
\format{A data frame with 15 observations on the following variable(s).\describe{
\item{\code{Effects}}{a character vector}
\item{\code{Value}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.21)}
\keyword{{datasets}}
