\name{Problem6.24}
\alias{Problem6.24}
\docType{data}
\title{Exercise 6.24}
\usage{data("Problem6.24")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Carbonation}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Speed}}{a numeric vector}
\item{\code{FillHeightDeviation}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.24)}
\keyword{{datasets}}
