\name{Problem6.25}
\alias{Problem6.25}
\docType{data}
\title{Exercise 6.25}
\usage{data("Problem6.25")}
\format{A data frame with 112 observations on the following variable(s).\describe{
\item{\code{LengthOfPutt}}{a numeric vector}
\item{\code{TypeOfPutter}}{a character vector}
\item{\code{BreakOfPutt}}{a character vector}
\item{\code{SlopeOfPutt}}{a character vector}
\item{\code{DistanceFromCup}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.25)}
\keyword{{datasets}}
