\name{Problem6.26}
\alias{Problem6.26}
\docType{data}
\title{Exercise 6.26}
\usage{data("Problem6.26")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{StdOrder}}{a numeric vector}
\item{\code{LaserPower}}{a numeric vector}
\item{\code{PulseFrequency}}{a numeric vector}
\item{\code{CellSize}}{a numeric vector}
\item{\code{WritingSpeed}}{a numeric vector}
\item{\code{Uec}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.26)}
\keyword{{datasets}}
