\name{Problem6.27}
\alias{Problem6.27}
\docType{data}
\title{Exercise 6.27}
\usage{data("Problem6.27")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{LaserPower}}{a numeric vector}
\item{\code{PulseFrequency}}{a numeric vector}
\item{\code{CellSize}}{a numeric vector}
\item{\code{WritingSpeed}}{a numeric vector}
\item{\code{Uec}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.27)}
\keyword{{datasets}}
