\name{Problem6.28}
\alias{Problem6.28}
\docType{data}
\title{Exercise 6.28}
\usage{data("Problem6.28")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Class}}{a character vector}
\item{\code{Type}}{a character vector}
\item{\code{Price}}{a numeric vector}
\item{\code{NumberOfOrders}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.28)}
\keyword{{datasets}}
