\name{Problem6.29}
\alias{Problem6.29}
\docType{data}
\title{Exercise 6.29}
\usage{data("Problem6.29")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Temperature}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Concentration}}{a numeric vector}
\item{\code{StirringRate}}{a numeric vector}
\item{\code{FiltrationRate}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.29)}
\keyword{{datasets}}
