\name{Problem6.31}
\alias{Problem6.31}
\docType{data}
\title{Exercise 6.31}
\usage{data("Problem6.31")}
\format{A data frame with 36 observations on the following variable(s).\describe{
\item{\code{Aperture}}{a character vector}
\item{\code{ExposureTime}}{a numeric vector}
\item{\code{DevelopTime}}{a numeric vector}
\item{\code{MaskDimension}}{a character vector}
\item{\code{EtchTime}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.31)}
\keyword{{datasets}}
