\name{Problem6.32}
\alias{Problem6.32}
\docType{data}
\title{Exercise 6.32}
\usage{data("Problem6.32")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Time}}{a numeric vector}
\item{\code{Concentration}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.32)}
\keyword{{datasets}}
