\name{Problem6.33}
\alias{Problem6.33}
\docType{data}
\title{Exercise 6.33}
\usage{data("Problem6.33")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Time}}{a numeric vector}
\item{\code{Concentration}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.33)}
\keyword{{datasets}}
