\name{Problem6.34}
\alias{Problem6.34}
\docType{data}
\title{Exercise 6.34}
\usage{data("Problem6.34")}
\format{A data frame with 64 observations on the following variable(s).\describe{
\item{\code{PanMaterial}}{a character vector}
\item{\code{StirringMethod}}{a character vector}
\item{\code{MixBrand}}{a character vector}
\item{\code{Scrumptiousness}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.34)}
\keyword{{datasets}}
