\name{Problem6.35}
\alias{Problem6.35}
\docType{data}
\title{Exercise 6.35}
\usage{data("Problem6.35")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{Temperature}}{a numeric vector}
\item{\code{CatalystConcentration}}{a numeric vector}
\item{\code{Time}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{MolecularWeight}}{a numeric vector}
\item{\code{Viscosity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.35)}
\keyword{{datasets}}
