\name{Problem6.39}
\alias{Problem6.39}
\docType{data}
\title{Exercise 6.39}
\usage{data("Problem6.39")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{ToolAngle}}{a numeric vector}
\item{\code{Viscosity}}{a numeric vector}
\item{\code{FeetRate}}{a numeric vector}
\item{\code{CuttingFluid}}{a character vector}
\item{\code{SurfaceRoughness}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.39)}
\keyword{{datasets}}
