\name{Problem6.40}
\alias{Problem6.40}
\docType{data}
\title{Exercise 6.40}
\usage{data("Problem6.40")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{D}}{a numeric vector}
\item{\code{Resistivity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.40)}
\keyword{{datasets}}
