\name{Problem6.42}
\alias{Problem6.42}
\docType{data}
\title{Exercise 6.42}
\usage{data("Problem6.42")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{D}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.42)}
\keyword{{datasets}}
