\name{Problem6.45}
\alias{Problem6.45}
\docType{data}
\title{Exercise 6.45}
\usage{data("Problem6.45")}
\format{A data frame with 22 observations on the following variable(s).\describe{
\item{\code{Glucose}}{a numeric vector}
\item{\code{Nh4no3}}{a numeric vector}
\item{\code{Feso4}}{a numeric vector}
\item{\code{Mnso4}}{a numeric vector}
\item{\code{Y}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.45)}
\keyword{{datasets}}
