\name{Problem6.46}
\alias{Problem6.46}
\docType{data}
\title{Exercise 6.46}
\usage{data("Problem6.46")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Apatite}}{a numeric vector}
\item{\code{Ph}}{a numeric vector}
\item{\code{Pb}}{a numeric vector}
\item{\code{FishbonePb}}{a numeric vector}
\item{\code{FishbonePh}}{a numeric vector}
\item{\code{HydroxyapatitePb}}{a numeric vector}
\item{\code{HydroxyapatitePh}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.46)}
\keyword{{datasets}}
