\name{Problem6.46i}
\alias{Problem6.46i}
\docType{data}
\title{Exercise 6.46i}
\usage{data("Problem6.46i")}
\format{A data frame with 32 observations on the following variable(s).\describe{
\item{\code{Apatite}}{a numeric vector}
\item{\code{Ph}}{a numeric vector}
\item{\code{Pb}}{a numeric vector}
\item{\code{Type}}{a character vector}
\item{\code{PbResponse}}{a numeric vector}
\item{\code{PhResponse}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.46i)}
\keyword{{datasets}}
