\name{Problem6.51}
\alias{Problem6.51}
\docType{data}
\title{Exercise 6.51}
\usage{data("Problem6.51")}
\format{A data frame with 32 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{D}}{a numeric vector}
\item{\code{E}}{a numeric vector}
\item{\code{Y}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.51)}
\keyword{{datasets}}
