\name{Problem6.52}
\alias{Problem6.52}
\docType{data}
\title{Exercise 6.52}
\usage{data("Problem6.52")}
\format{A data frame with 12 observations on the following variable(s).\describe{
\item{\code{Run}}{a numeric vector}
\item{\code{Temp}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Conc}}{a numeric vector}
\item{\code{X1}}{a numeric vector}
\item{\code{X2}}{a numeric vector}
\item{\code{X3}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.52)}
\keyword{{datasets}}
