\name{Problem6.9}
\alias{Problem6.9}
\docType{data}
\title{Exercise 6.9}
\usage{data("Problem6.9")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{BitSize}}{a numeric vector}
\item{\code{CuttingSpeed}}{a numeric vector}
\item{\code{Vibration}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem6.9)}
\keyword{{datasets}}
