\name{Problem7.10}
\alias{Problem7.10}
\docType{data}
\title{Exercise 7.10}
\usage{data("Problem7.10")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{Carbonation}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Speed}}{a numeric vector}
\item{\code{FillHeightDeviation}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem7.10)}
\keyword{{datasets}}
