\name{Problem7.12}
\alias{Problem7.12}
\docType{data}
\title{Exercise 7.12}
\usage{data("Problem7.12")}
\format{A data frame with 112 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{LengthOfPutt}}{a numeric vector}
\item{\code{TypeOfPutter}}{a character vector}
\item{\code{BreakOfPutt}}{a character vector}
\item{\code{SlopeOfPutt}}{a character vector}
\item{\code{DistanceFromCup}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem7.12)}
\keyword{{datasets}}
