\name{Problem7.13}
\alias{Problem7.13}
\docType{data}
\title{Exercise 7.13}
\usage{data("Problem7.13")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{LaserPower}}{a numeric vector}
\item{\code{PulseFrequency}}{a numeric vector}
\item{\code{CellSize}}{a numeric vector}
\item{\code{WritingSpeed}}{a numeric vector}
\item{\code{Uec}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem7.13)}
\keyword{{datasets}}
