\name{Problem7.14}
\alias{Problem7.14}
\docType{data}
\title{Exercise 7.14}
\usage{data("Problem7.14")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{Class}}{a character vector}
\item{\code{Type}}{a character vector}
\item{\code{Price}}{a numeric vector}
\item{\code{NumberOfOrders}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem7.14)}
\keyword{{datasets}}
