\name{Problem7.15}
\alias{Problem7.15}
\docType{data}
\title{Exercise 7.15}
\usage{data("Problem7.15")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{AcidStrength}}{a numeric vector}
\item{\code{ReactionTime}}{a numeric vector}
\item{\code{AmountOfAcid}}{a numeric vector}
\item{\code{ReactionTemperature}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem7.15)}
\keyword{{datasets}}
