\name{Problem7.18}
\alias{Problem7.18}
\docType{data}
\title{Exercise 7.18}
\usage{data("Problem7.18")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{Glucose}}{a numeric vector}
\item{\code{Nh4no3}}{a numeric vector}
\item{\code{Feso4}}{a numeric vector}
\item{\code{Mnso4}}{a numeric vector}
\item{\code{Y}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem7.18)}
\keyword{{datasets}}
