\name{Problem7.2}
\alias{Problem7.2}
\docType{data}
\title{Exercise 7.2}
\usage{data("Problem7.2")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{BitSize}}{a numeric vector}
\item{\code{CuttingSpeed}}{a numeric vector}
\item{\code{Vibration}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem7.2)}
\keyword{{datasets}}
