\name{Problem7.23}
\alias{Problem7.23}
\docType{data}
\title{Exercise 7.23}
\usage{data("Problem7.23")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Concentration}}{a numeric vector}
\item{\code{StirringRate}}{a numeric vector}
\item{\code{FiltrationRate}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem7.23)}
\keyword{{datasets}}
