\name{Problem7.25}
\alias{Problem7.25}
\docType{data}
\title{Exercise 7.25}
\usage{data("Problem7.25")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{CuttingSpeed}}{a numeric vector}
\item{\code{ToolGeometry}}{a numeric vector}
\item{\code{CuttingAngle}}{a numeric vector}
\item{\code{LifeHours}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem7.25)}
\keyword{{datasets}}
