\name{Problem7.3}
\alias{Problem7.3}
\docType{data}
\title{Exercise 7.3}
\usage{data("Problem7.3")}
\format{A data frame with 32 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{PourTemperature}}{a numeric vector}
\item{\code{TitaniumContent}}{a numeric vector}
\item{\code{HeatTreatMethod}}{a numeric vector}
\item{\code{GrainRefiner}}{a numeric vector}
\item{\code{CrackLength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem7.3)}
\keyword{{datasets}}
