\name{Problem7.8}
\alias{Problem7.8}
\docType{data}
\title{Exercise 7.8}
\usage{data("Problem7.8")}
\format{A data frame with 32 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{Aperture}}{a character vector}
\item{\code{ExposureTime}}{a numeric vector}
\item{\code{DevelopTime}}{a numeric vector}
\item{\code{MaskDimension}}{a character vector}
\item{\code{EtchTime}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem7.8)}
\keyword{{datasets}}
