\name{Problem8.1}
\alias{Problem8.1}
\docType{data}
\title{Exercise 8.1}
\usage{data("Problem8.1")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{D}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.1)}
\keyword{{datasets}}
