\name{Problem8.10}
\alias{Problem8.10}
\docType{data}
\title{Exercise 8.10}
\usage{data("Problem8.10")}
\format{A data frame with 48 observations on the following variable(s).\describe{
\item{\code{FurnaceTemp}}{a numeric vector}
\item{\code{HeatingTime}}{a numeric vector}
\item{\code{TransferTime}}{a numeric vector}
\item{\code{HoldTime}}{a numeric vector}
\item{\code{QuenchTemp}}{a numeric vector}
\item{\code{FreeHeight}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.10)}
\keyword{{datasets}}
