\name{Problem8.10c}
\alias{Problem8.10c}
\docType{data}
\title{Exercise 8.10c}
\usage{data("Problem8.10c")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{FurnaceTemp}}{a numeric vector}
\item{\code{HeatingTime}}{a numeric vector}
\item{\code{TransferTime}}{a numeric vector}
\item{\code{HoldTime}}{a numeric vector}
\item{\code{QuenchTemp}}{a numeric vector}
\item{\code{Range}}{a numeric vector}
\item{\code{Stdev}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.10c)}
\keyword{{datasets}}
