\name{Problem8.11}
\alias{Problem8.11}
\docType{data}
\title{Exercise 8.11}
\usage{data("Problem8.11")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{Condensation}}{a numeric vector}
\item{\code{Material1}}{a numeric vector}
\item{\code{Solvent}}{a numeric vector}
\item{\code{Time}}{a numeric vector}
\item{\code{Material2}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.11)}
\keyword{{datasets}}
