\name{Problem8.14}
\alias{Problem8.14}
\docType{data}
\title{Exercise 8.14}
\usage{data("Problem8.14")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{Aperture}}{a character vector}
\item{\code{ExposureTime}}{a numeric vector}
\item{\code{DevelopTime}}{a numeric vector}
\item{\code{MaskDimension}}{a character vector}
\item{\code{EtchTime}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.14)}
\keyword{{datasets}}
