\name{Problem8.15}
\alias{Problem8.15}
\docType{data}
\title{Exercise 8.15}
\usage{data("Problem8.15")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{Time}}{a numeric vector}
\item{\code{Concentration}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.15)}
\keyword{{datasets}}
