\name{Problem8.17}
\alias{Problem8.17}
\docType{data}
\title{Exercise 8.17}
\usage{data("Problem8.17")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{Temperature}}{a numeric vector}
\item{\code{Pressure}}{a numeric vector}
\item{\code{FiltrationRate}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.17)}
\keyword{{datasets}}
