\name{Problem8.2}
\alias{Problem8.2}
\docType{data}
\title{Exercise 8.2}
\usage{data("Problem8.2")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{PourTemp}}{a numeric vector}
\item{\code{TitaniumContent}}{a numeric vector}
\item{\code{HeatTreatMethod}}{a numeric vector}
\item{\code{GrainRefiner}}{a numeric vector}
\item{\code{CrackLength}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.2)}
\keyword{{datasets}}
