\name{Problem8.27}
\alias{Problem8.27}
\docType{data}
\title{Exercise 8.27}
\usage{data("Problem8.27")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{Pitch/finesRatio}}{a numeric vector}
\item{\code{PackType}}{a character vector}
\item{\code{PackTemp}}{a numeric vector}
\item{\code{FlueLocation}}{a character vector}
\item{\code{PitTemp}}{a numeric vector}
\item{\code{DelayTime}}{a numeric vector}
\item{\code{Weight}}{a numeric vector}
\item{\code{Range}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.27)}
\keyword{{datasets}}
