\name{Problem8.28}
\alias{Problem8.28}
\docType{data}
\title{Exercise 8.28}
\usage{data("Problem8.28")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{LamTemp}}{a numeric vector}
\item{\code{LamTime}}{a numeric vector}
\item{\code{LamPres}}{a numeric vector}
\item{\code{FireTemp}}{a numeric vector}
\item{\code{FireTime}}{a numeric vector}
\item{\code{FireDp}}{a numeric vector}
\item{\code{CamberAvg}}{a numeric vector}
\item{\code{CamberStdev}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.28)}
\keyword{{datasets}}
