\name{Problem8.29}
\alias{Problem8.29}
\docType{data}
\title{Exercise 8.29}
\usage{data("Problem8.29")}
\format{A data frame with 32 observations on the following variable(s).\describe{
\item{\code{Volume}}{a numeric vector}
\item{\code{Batch}}{a character vector}
\item{\code{Time}}{a numeric vector}
\item{\code{Speed}}{a numeric vector}
\item{\code{Acc}}{a numeric vector}
\item{\code{Cover}}{a character vector}
\item{\code{ThickAvg}}{a numeric vector}
\item{\code{Range}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.29)}
\keyword{{datasets}}
