\name{Problem8.2c}
\alias{Problem8.2c}
\docType{data}
\title{Exercise 8.2c}
\usage{data("Problem8.2c")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{OrderQuantity}}{a numeric vector}
\item{\code{Re-orderPoint}}{a numeric vector}
\item{\code{SetupCost}}{a numeric vector}
\item{\code{BackorderCost}}{a numeric vector}
\item{\code{CarryingCost}}{a numeric vector}
\item{\code{AvgAnnualCost}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.2c)}
\keyword{{datasets}}
