\name{Problem8.3}
\alias{Problem8.3}
\docType{data}
\title{Exercise 8.3}
\usage{data("Problem8.3")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{Gap}}{a numeric vector}
\item{\code{C2f6Flow}}{a numeric vector}
\item{\code{Power}}{a numeric vector}
\item{\code{EtchRate}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.3)}
\keyword{{datasets}}
