\name{Problem8.30d}
\alias{Problem8.30d}
\docType{data}
\title{Exercise 8.30d}
\usage{data("Problem8.30d")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{A}}{a character vector}
\item{\code{B}}{a character vector}
\item{\code{C}}{a character vector}
\item{\code{D}}{a character vector}
\item{\code{E}}{a character vector}
\item{\code{F}}{a character vector}
\item{\code{G}}{a character vector}
\item{\code{H}}{a character vector}
\item{\code{TasteAvg}}{a numeric vector}
\item{\code{TasteStdev}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.30d)}
\keyword{{datasets}}
