\name{Problem8.31}
\alias{Problem8.31}
\docType{data}
\title{Exercise 8.31}
\usage{data("Problem8.31")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{AcidStrength}}{a numeric vector}
\item{\code{ReactionTime}}{a numeric vector}
\item{\code{AmountOfAcid}}{a numeric vector}
\item{\code{ReactionTemperature}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.31)}
\keyword{{datasets}}
