\name{Problem8.32}
\alias{Problem8.32}
\docType{data}
\title{Exercise 8.32}
\usage{data("Problem8.32")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{D}}{a numeric vector}
\item{\code{E}}{a numeric vector}
\item{\code{Y}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.32)}
\keyword{{datasets}}
