\name{Problem8.33}
\alias{Problem8.33}
\docType{data}
\title{Exercise 8.33}
\usage{data("Problem8.33")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{Glucose}}{a numeric vector}
\item{\code{Nh4no3}}{a numeric vector}
\item{\code{Feso4}}{a numeric vector}
\item{\code{Mnso4}}{a numeric vector}
\item{\code{Y}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.33)}
\keyword{{datasets}}
