\name{Problem8.34}
\alias{Problem8.34}
\docType{data}
\title{Exercise 8.34}
\usage{data("Problem8.34")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Apatite}}{a numeric vector}
\item{\code{Ph}}{a numeric vector}
\item{\code{Pb}}{a numeric vector}
\item{\code{Type}}{a character vector}
\item{\code{PbResponseMm}}{a numeric vector}
\item{\code{PhResponse}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.34)}
\keyword{{datasets}}
