\name{Problem8.35}
\alias{Problem8.35}
\docType{data}
\title{Exercise 8.35}
\usage{data("Problem8.35")}
\format{A data frame with 17 observations on the following variable(s).\describe{
\item{\code{P}}{a numeric vector}
\item{\code{T}}{a numeric vector}
\item{\code{S}}{a numeric vector}
\item{\code{D}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{Resolution}}{a numeric vector}
\item{\code{Sensitivity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.35)}
\keyword{{datasets}}
