\name{Problem8.36}
\alias{Problem8.36}
\docType{data}
\title{Exercise 8.36}
\usage{data("Problem8.36")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{Sih2cl2}}{a numeric vector}
\item{\code{Geh4}}{a numeric vector}
\item{\code{Hcl}}{a numeric vector}
\item{\code{B2h6}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{SiCapThickness}}{a numeric vector}
\item{\code{SigeThickness}}{a numeric vector}
\item{\code{GeConcentration}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.36)}
\keyword{{datasets}}
