\name{Problem8.37}
\alias{Problem8.37}
\docType{data}
\title{Exercise 8.37}
\usage{data("Problem8.37")}
\format{A data frame with 32 observations on the following variable(s).\describe{
\item{\code{SqueegeePressure}}{a numeric vector}
\item{\code{PrintingSpeed}}{a numeric vector}
\item{\code{SqueegeeAngle}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Viscosity}}{a numeric vector}
\item{\code{CleaningInterval}}{a numeric vector}
\item{\code{SeparationSpeed}}{a numeric vector}
\item{\code{RelativeHumidity}}{a numeric vector}
\item{\code{Pvm}}{a numeric vector}
\item{\code{Npu}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.37)}
\keyword{{datasets}}
