\name{Problem8.38}
\alias{Problem8.38}
\docType{data}
\title{Exercise 8.38}
\usage{data("Problem8.38")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{EnvelopeTeaser}}{a character vector}
\item{\code{ReturnAddress}}{a character vector}
\item{\code{OfficialInk-stamp}}{a character vector}
\item{\code{Postage}}{a character vector}
\item{\code{AdditionalGraphics}}{a character vector}
\item{\code{PriceGraphic}}{a character vector}
\item{\code{Sticker}}{a character vector}
\item{\code{Personalizeletter}}{a character vector}
\item{\code{CopyMessage}}{a character vector}
\item{\code{LetterHeadline}}{a character vector}
\item{\code{ListOfBenefits}}{a character vector}
\item{\code{PostscriptOnLetter}}{a character vector}
\item{\code{Signature}}{a character vector}
\item{\code{ProductSelection}}{a character vector}
\item{\code{ValueOfFreeGift}}{a character vector}
\item{\code{ReplyEnvelope}}{a character vector}
\item{\code{InformationOnBuckslip}}{a character vector}
\item{\code{SecondBuckslip}}{a character vector}
\item{\code{InterestRate}}{a character vector}
\item{\code{Orders}}{a numeric vector}
\item{\code{ResponseRate}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.38)}
\keyword{{datasets}}
