\name{Problem8.38e}
\alias{Problem8.38e}
\docType{data}
\title{Exercise 8.38e}
\usage{data("Problem8.38e")}
\format{A data frame with 20 observations on the following variable(s).\describe{
\item{\code{Sticker}}{a character vector}
\item{\code{SecondBuckslip}}{a character vector}
\item{\code{InterestRate}}{a character vector}
\item{\code{Orders}}{a numeric vector}
\item{\code{ResponseRate}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.38e)}
\keyword{{datasets}}
