\name{Problem8.4}
\alias{Problem8.4}
\docType{data}
\title{Exercise 8.4}
\usage{data("Problem8.4")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{Aperture}}{a character vector}
\item{\code{ExposureTime}}{a numeric vector}
\item{\code{DevelopTime}}{a numeric vector}
\item{\code{MaskDimension}}{a character vector}
\item{\code{EtchTime}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.4)}
\keyword{{datasets}}
