\name{Problem8.46}
\alias{Problem8.46}
\docType{data}
\title{Exercise 8.46}
\usage{data("Problem8.46")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Run}}{a numeric vector}
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{D}}{a numeric vector}
\item{\code{E}}{a numeric vector}
\item{\code{Y}}{a numeric vector}
\item{\code{Block}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.46)}
\keyword{{datasets}}
