\name{Problem8.49}
\alias{Problem8.49}
\docType{data}
\title{Exercise 8.49}
\usage{data("Problem8.49")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Pressure}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Moisture}}{a numeric vector}
\item{\code{Flow}}{a numeric vector}
\item{\code{ParticleSize}}{a numeric vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.49)}
\keyword{{datasets}}
