\name{Problem8.50}
\alias{Problem8.50}
\docType{data}
\title{Exercise 8.50}
\usage{data("Problem8.50")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{A}}{a numeric vector}
\item{\code{B}}{a numeric vector}
\item{\code{C}}{a numeric vector}
\item{\code{D}}{a numeric vector}
\item{\code{E}}{a numeric vector}
\item{\code{F}}{a numeric vector}
\item{\code{G}}{a numeric vector}
\item{\code{H}}{a numeric vector}
\item{\code{J}}{a numeric vector}
\item{\code{K}}{a numeric vector}
\item{\code{P}}{a numeric vector}
\item{\code{Arcsin}}{a numeric vector}
\item{\code{FAndT}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.50)}
\keyword{{datasets}}
