\name{Problem8.54}
\alias{Problem8.54}
\docType{data}
\title{Exercise 8.54}
\usage{data("Problem8.54")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{MixerPorts}}{a character vector}
\item{\code{Temperature}}{a character vector}
\item{\code{MixingTime}}{a numeric vector}
\item{\code{BatchWeight}}{a numeric vector}
\item{\code{Delays}}{a numeric vector}
\item{\code{StdDev}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.54)}
\keyword{{datasets}}
