\name{Problem8.7}
\alias{Problem8.7}
\docType{data}
\title{Exercise 8.7}
\usage{data("Problem8.7")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Block}}{a numeric vector}
\item{\code{AnnualFee}}{a character vector}
\item{\code{Account-openingFee}}{a character vector}
\item{\code{InitialInterest}}{a character vector}
\item{\code{Long-termInterestRate}}{a character vector}
\item{\code{ResponseRate}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.7)}
\keyword{{datasets}}
