\name{Problem8.9}
\alias{Problem8.9}
\docType{data}
\title{Exercise 8.9}
\usage{data("Problem8.9")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Solvent}}{a numeric vector}
\item{\code{Catalyst}}{a numeric vector}
\item{\code{Temperature}}{a numeric vector}
\item{\code{Purity}}{a numeric vector}
\item{\code{ReactantPh}}{a numeric vector}
\item{\code{Color}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem8.9)}
\keyword{{datasets}}
