\name{Problem9.1}
\alias{Problem9.1}
\docType{data}
\title{Exercise 9.1}
\usage{data("Problem9.1")}
\format{A data frame with 36 observations on the following variable(s).\describe{
\item{\code{Strength}}{a character vector}
\item{\code{Time}}{a character vector}
\item{\code{Density}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem9.1)}
\keyword{{datasets}}
