\name{Problem9.22}
\alias{Problem9.22}
\docType{data}
\title{Exercise 9.22}
\usage{data("Problem9.22")}
\format{A data frame with 16 observations on the following variable(s).\describe{
\item{\code{Clone}}{a character vector}
\item{\code{BerrySize}}{a character vector}
\item{\code{FermTemp1}}{a character vector}
\item{\code{FermTemp2}}{a character vector}
\item{\code{WholeBerry}}{a character vector}
\item{\code{MacerTime}}{a numeric vector}
\item{\code{Yeast}}{a character vector}
\item{\code{Oak}}{a character vector}
\item{\code{Rank}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem9.22)}
\keyword{{datasets}}
