\name{Problem9.23b}
\alias{Problem9.23b}
\docType{data}
\title{Exercise 9.23b}
\usage{data("Problem9.23b")}
\format{A data frame with 95 observations on the following variable(s).\describe{
\item{\code{Time}}{a character vector}
\item{\code{Process}}{a character vector}
\item{\code{Machine}}{a character vector}
\item{\code{Length}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem9.23b)}
\keyword{{datasets}}
