\name{Problem9.23d}
\alias{Problem9.23d}
\docType{data}
\title{Exercise 9.23d}
\usage{data("Problem9.23d")}
\format{A data frame with 24 observations on the following variable(s).\describe{
\item{\code{Time}}{a character vector}
\item{\code{Process}}{a character vector}
\item{\code{Machine}}{a character vector}
\item{\code{Avg}}{a numeric vector}
\item{\code{Log(var)}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem9.23d)}
\keyword{{datasets}}
