\name{Problem9.23e}
\alias{Problem9.23e}
\docType{data}
\title{Exercise 9.23e}
\usage{data("Problem9.23e")}
\format{A data frame with 8 observations on the following variable(s).\describe{
\item{\code{Process}}{a character vector}
\item{\code{Machine}}{a character vector}
\item{\code{Avg}}{a numeric vector}
\item{\code{Var}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem9.23e)}
\keyword{{datasets}}
