\name{Problem9.28}
\alias{Problem9.28}
\docType{data}
\title{Exercise 9.28}
\usage{data("Problem9.28")}
\format{A data frame with 27 observations on the following variable(s).\describe{
\item{\code{PctCarbon}}{a numeric vector}
\item{\code{PctNitrogen}}{a numeric vector}
\item{\code{Temp}}{a numeric vector}
\item{\code{Activity}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem9.28)}
\keyword{{datasets}}
