\name{Problem9.4}
\alias{Problem9.4}
\docType{data}
\title{Exercise 9.4}
\usage{data("Problem9.4")}
\format{A data frame with 54 observations on the following variable(s).\describe{
\item{\code{Brand}}{a character vector}
\item{\code{Strength}}{a character vector}
\item{\code{Dog}}{a character vector}
\item{\code{EnzymeLevel}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem9.4)}
\keyword{{datasets}}
