\name{Problem9.6}
\alias{Problem9.6}
\docType{data}
\title{Exercise 9.6}
\usage{data("Problem9.6")}
\format{A data frame with 18 observations on the following variable(s).\describe{
\item{\code{Pressure}}{a character vector}
\item{\code{Temperature}}{a character vector}
\item{\code{Yield}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem9.6)}
\keyword{{datasets}}
