\name{Problem9.9}
\alias{Problem9.9}
\docType{data}
\title{Exercise 9.9}
\usage{data("Problem9.9")}
\format{A data frame with 27 observations on the following variable(s).\describe{
\item{\code{Block}}{a character vector}
\item{\code{BottleType}}{a character vector}
\item{\code{ShelfType}}{a character vector}
\item{\code{Worker}}{a character vector}
\item{\code{Time}}{a numeric vector}
}}
\references{Montgomery, D.C.(2017, 10th ed.) \emph{Design and Analysis of Experiments}, Wiley, New York.}
\examples{data(Problem9.9)}
\keyword{{datasets}}
